Fix stream audio
The Sound(On|Off)Icon would turn off after the first track ends.
This happened thanks to code that handles audio tracks played for
a single ayah. We can identify the context where we don't want to
automatically disable via 'autoPlay'. Not a great fix but works.

We also top up the Timer count when sound is turned on. That helps
make sure the audio will finish before the timer.