audio: import & improve useAudio (via EasyQuran)
The basis for this code comes from EasyQuran, and it has
been improved and modified in the process. The 'useAudio'
hook now returns an abstraction built around HTMLAudioElement.

Rather than having many disconnected states, the abstraction
encapsulates everything to do with an audio element in a single
object and sets the stage for more refactors in the future -
Inshallah.