audio: fixes for mobile, and other improvements
This change fixes buggy behavior on mobile (tested on iOS), and it
also synchronizes the play|pause icons to the playback of audio as
well. Previously, you could pause the stream and the audio would
continue playing regardless. With this change, the audio is paused
and it can also be resumed from the point of pause. The timer also
picks up from where it left off.